// apps/admin/src/components/schedule/types.ts
export enum Preset {
  // 운영 일정
  OPERATION_START = "OPERATION_START",
  OPERATION_PRESTART = "OPERATION_PRESTART",
  OPERATION_MAINTENANCE_START = "OPERATION_MAINTENANCE_START",
  OPERATION_MAINTENANCE_END = "OPERATION_MAINTENANCE_END",

  // 모집 일정
  APPLICATION_START = "APPLICATION_START",
  APPLICATION_END = "APPLICATION_END",

  ETC = "ETC",
}

export const PresetTitle = {
  // 운영 일정
  [Preset.OPERATION_START]: "오픈",
  [Preset.OPERATION_PRESTART]: "가오픈",
  [Preset.OPERATION_MAINTENANCE_START]: "점검 시작",
  [Preset.OPERATION_MAINTENANCE_END]: "점검 종료",

  // 모집 일정
  [Preset.APPLICATION_START]: "모집 시작",
  [Preset.APPLICATION_END]: "모집 종료",
}